Fix some errors in IT sample loading and decompression

+++ src/ca/mrb0/hydrocitee/it/ITSampleData.java
@@ -42,9 +42,9 @@ public class ITSampleData {
             outData = Lists.newArrayListWithCapacity(dataLengthSamples);
             for(int i = 0; i < dataLengthSamples; i++) {
                 if (is16Bit) {
-                    outData.add(((0xff & data[offs + 2 * i + 1]) << 8) | (0xff & data[offs + 2 * i])); 
+                    outData.add((int)(short)(((0xff & data[offs + 2 * i + 1]) << 8) | (0xff & data[offs + 2 * i]))); 
                 } else {
-                    outData.add(0xff & data[offs + i]);
+                    outData.add((int)(byte)(0xff & data[offs + i]));
                 }
             }
         }
